setup namespaces [not tested]
