+++ src/Velo.class.php
@@ -1,4 +1,7 @@
 <?php
+
+namespace Irail\BikesResource;
+
 /**
  *
  * @package packages/Bikes
@@ -6,10 +9,7 @@
  * @license AGPLv3
  * @author Pieter Colpaert <pieter aŧ iRail.be>
  */
-
-include_once('gpoint.php');
-
-class BikesVillo extends AReader
+class BikesVelo extends AReader
 {
 
     public function __construct($package, $resource, $RESTparameters)
@@ -51,32 +51,24 @@ class BikesVillo extends AReader
 
     public function read()
     {
-        $data = TDT::HttpRequest("http://www.mobielbrussel.irisnet.be/villo/json/");
+        $data = TDT::HttpRequest("http://ClearChannelBikes:WbRpWHk6Ur@m.velo-antwerpen.be/data/stations.json");
         $decoded = json_decode($data->data);
         //todo: convert to wished format
 
         $result = array();
         $gpoint = new gPoint();
 
-        foreach ($decoded->features as $feature) {
+        foreach ($decoded as $sourceStation) {
             $station = new Object();
 
-            $station->name = $feature->properties->NAME;
-            $station->freebikes = $feature->properties->FREEBK;
-            $station->freespots = $feature->properties->FREEBS;
-            $station->state = $feature->properties->STATE;
+            $station->name = $sourceStation->name;
+            $station->freebikes = $sourceStation->bikes;
+            $station->freespots = $sourceStation->slots;
+            $station->state = $sourceStation->status;
+            $station->latitude = $sourceStation->latitude;
+            $station->longitude = $sourceStation->longitude;
 
-            // Configure the gPoint library to use the Lambert Projection for Belgium
-            $gpoint->configLambertProjection(150000.013, 5400088.438, 4.367487, 90, 49.833333, 51.166666);
-            $x = $feature->geometry->coordinates[0];
-            $y = $feature->geometry->coordinates[1];
-
-            $gpoint->setLambert($x, $y);
-            // Convert the Lambert Coordinates to Latitude and Longitude (using the gPoint Library)
-            $gpoint->convertLCCtoLL();
-
-            $station->latitude = $gpoint->lat;
-            $station->longitude = $gpoint->long;
+            $gpoint->setLongLat($station->longitude, $station->latitude);
 
             if ($this->lat != null && $this->long != null) {
                 $station->distance = $gpoint->distanceFrom($this->long, $this->lat);
@@ -102,6 +94,6 @@ class BikesVillo extends AReader
 
     public static function getDoc()
     {
-        return "This resource contains dynamic information about the availability of bikes in Brussels";
+        return "This resource contains dynamic information about the availability of bikes in Antwerp";
     }
 }
