+++ src/resources.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Irail\BikesResource;
+
 /* Copyright (C) 2012 by iRail vzw/asbl
  *
  * Author: Maarten Cautreels <maarten at flatturtle.com>
